// frontend/src/types/product.ts
// src/types/product.ts
export interface Product {
  id_produto: string;
  nome_produto: string;
  categoria_produto: string;
  peso_produto_gramas: number;
  total_vendas: number;
  media_avaliacoes: number;
  comprimento_centimetros: number;
  altura_centimetros: number;
  largura_centimetros: number;
}